requirements/base.txt
Django==1.9.1
GitPython==1.0.1
colorama==0.3.6
gitdb==0.6.4
keyring==7.3
psycopg2==2.6.1
pyftpsync==1.0.3
smmap==0.9.0
